+++ src/main/java/com/quasiris/qsf/pipeline/filter/mapper/DefaultFacetKeyMapper.java
@@ -4,7 +4,7 @@ public class DefaultFacetKeyMapper implements FacetKeyMapper {
 
 
     @Override
-    public String map(String value) {
+    public Object map(Object value) {
         return value;
     }
 }