+++ src/main/java/com/quasiris/qsf/pipeline/filter/mapper/FacetKeyMapper.java
@@ -2,5 +2,5 @@ package com.quasiris.qsf.pipeline.filter.mapper;
 
 public interface FacetKeyMapper {
 
-    String map(String value);
+    Object map(Object value);
 }