+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes qsf-integration project will be documented in this file.
 
 ### Changed
 - use java http client as default in elastic client and elastic filter
+- use Object for facet value to support non string values like integer, double, ...
 
 ## [6.10]
 ### Added